backend/requirements.txt
# GARMF-Mpox Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ML Core
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

# Deep Learning
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0

# Imbalanced Learning
imbalanced-learn==0.11.0

# XGBoost & LightGBM
xgboost==2.0.2
lightgbm==4.1.0

# Explainability
shap==0.43.0
opencv-python==4.8.1.78

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Logging & Monitoring
structlog==23.2.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
